remove unnessesary variables
