+++ Makefile
@@ -1,7 +1,7 @@
 COMPILER = gcc
 TARGET = kvmm
 CFLAGS = -pthread
-OBJECTS = main.o util.o blk.o io.o lapic.o interrupt.o ioapic.o uart.o vm.o vcpu.o
+OBJECTS = main.o util.o blk.o io.o lapic.o interrupt.o ioapic.o uart.o vm.o vcpu.o mmio.o
 
 all: 
 	make $(TARGET)
