+++ interrupt.c
@@ -1,29 +1,3 @@
-#include "interrupt.h"
-#include "util.h"
-
-void enq_irr(struct irr_queue *irr, int value) {
-    irr->buff[irr->last] = value;
-    irr->last++;
-}
-
-int deq_irr(struct irr_queue *irr) {
-   int out = irr->buff[0];
-   for (int i = 0; i <= irr->last; i++) {
-       irr->buff[i] = irr->buff[i+1];
-   }
-   irr->last--;
-}  
-
-void inject_interrupt(int vcpufd, int irq) {
-    struct kvm_interrupt *intr = malloc(sizeof(struct kvm_interrupt));
-    intr->irq = irq;
-    
-    
-    if (ioctl(vcpufd, KVM_INTERRUPT, intr) < 0)
-        error("KVM_INTERRUPT");
-}
-
-
 // struct interrupt_buffer *init_irr_buff() {
 //     struct interrupt_buffer *irr_buff = malloc(sizeof(struct interrupt_buffer));
 //     if (irr_buff == NULL)
