+++ ioapic.h
@@ -33,4 +33,5 @@ struct ioapic {
 #define IOAPIC_BASE 0xfec00000
 #define IOAPIC_REDRTB_BASE (IOAPIC_BASE + 0x10)
 
+void init_ioapic();
 void emulate_ioapicw(struct vcpu *vcpu);
