+++ mmio.h
@@ -1,6 +1,10 @@
 #pragma once
 
 #include <linux/kvm.h>
+#include "vcpu.h"
+#include "vm.h"
+#include "lapic.h"
+#include "ioapic.h"
 
 struct mmio {
     __u64 phys_addr;
@@ -9,3 +13,4 @@ struct mmio {
 	__u8  is_write;
 };
 
+void emulate_mmio(struct vcpu *vcpu);
